websockets/orderbook/requirements.txt
websocket-client
pyTelegramBotAPI
numpy
SQLAlchemy
psycopg2-binary